requirements.txt
flask
numpy
faiss-cpu
PyMuPDF
requests
python-dotenv
flask-cors
pgvector
langchain
psycopg
pytesseract
psycopg[binary]
Pillow